Check if artist and track name are still now playing
Searching for the lyrics on the different lyrics providers might take 
some time and during that time the song might have already changed. In 
the past this resulted in the lyrics of the previous track being shown 
once loading had finished. Now it just quietly saves the lyrics without 
showing it.